requirements.txt
Flask
pika
boto3
pika
boto3
pylint
autopep8

pyyaml
